Cleaned code, added javadocs
